Uglify Javascript && Add PostCSS plugins
