+++ package.json
@@ -17,14 +17,17 @@
   },
   "devDependencies": {
     "bower": "^1.6.5",
+    "cssnano": "^3.3.2",
     "eslint": "^1.10.1",
     "gulp": "^3.9.0",
     "gulp-autoprefixer": "^3.1.0",
     "gulp-concat": "^2.6.0",
     "gulp-connect": "^2.2.0",
     "gulp-cssimport": "^2.1.1",
+    "gulp-mode": "^1.0.1",
     "gulp-postcss": "^6.0.1",
     "gulp-sass": "^2.0.4",
+    "gulp-uglify": "^1.5.1",
     "handlebars": "^4.0.1",
     "handlebars-layouts": "^3.1.2",
     "lodash": "^3.10.1",
@@ -46,10 +49,10 @@
     "minimist": "^1.2.0",
     "mocha": "^2.3.4",
     "moment": "^2.10.6",
+    "pixrem": "^3.0.0",
     "postcss-import": "^7.1.3",
     "slug": "^0.9.1",
     "surge": "^0.17.3",
-    "webpack": "^1.12.9",
     "zombie": "^4.2.1"
   },
   "dependencies": {
